refactor: extract HandleSetupDelegate from GlobalDragAdapter (534→334 lines)
Extract the 195-line setupHandles() method body into a package-private
HandleSetupDelegate class with a static setupHandles(DragAdapter) method.

Changes:
- Create HandleSetupDelegate.java with all handle creation logic
- Replace setupHandles() call with HandleSetupDelegate.setupHandles(this)
- Remove private setupHandles() method from GlobalDragAdapter
- Remove unused Color4f and ManipulationEvent/Criteria imports
- Retain Resizer import (still used in ResizeDragManipulator at line 173)
- Fix contract test path resolution (Maven runs from core/ide dir)
- Fix contract test: Resizer import is retained, not removed

All 1249 tests pass including 43 new extraction tests.

Closes #685
